add authonticatin and authorization by craeting user model and auth middleware

+++ controllers/memories.js
@@ -9,6 +9,17 @@ export const getMemories = async (req, res) => {
   }
 };
 
+export const getMemory = async (req, res) => {
+  const { id } = req.params;
+  if (!req.userId) return res.json({ message: "Unauthenticated.." });
+  try {
+    const memories = await Memory.findById(id);
+    res.status(200).json(memories);
+  } catch (error) {
+    res.status(404).json({ message: error.message });
+  }
+};
+
 export const createMemory = async (req, res) => {
   const memory = req.body;
   const newMemory = new Memory(memory);
@@ -22,41 +33,64 @@ export const createMemory = async (req, res) => {
 
 export const updateMemory = async (req, res) => {
   const { id } = req.params;
-  const memory = req.body;
+  const updateMemory = req.body;
+
+  if (!req.userId) return res.json({ message: "Unauthenticated.." });
 
   if (!mongoose.Type.ObjectId.isValid(id))
     return res.status(404).send("Now Memory with this Id");
 
-  const updateMemory = await Memory.findByIdAndUpdate(id, memory, {
-    new: true,
-  });
+  const memory = await Memory.findById(id);
 
-  res.json(updateMemory);
+  const updatedMemory;
+  if (String(memory.userId) === String(req.userId)) {
+    updatedMemory = await Memory.findByIdAndUpdate(id, updateMemory, {
+      new: true,
+    });
+  }
+  res.json(updatedMemory);
 };
 
 export const deleteMemory = async (req, res) => {
   const { id } = req.params;
 
+  if (!req.userId) return res.json({ message: "Unauthenticated.." });
+
   if (!mongoose.Type.ObjectId.isValid(id))
     return res.status(404).send("Now Memory with this Id");
 
-  await Memory.findByIdAndRemove(id);
+  const memory = await Memory.findById(id);
 
+  if (String(memory.userId) === String(req.userId)) {
+    await Memory.findByIdAndRemove(id);
+  }
   res.json({ message: "Memory deleted successfully" });
 };
 
 export const likeMemory = async (req, res) => {
   const { id } = req.params;
 
+  if (!req.userId) return res.json({ message: "Unauthenticated.." });
+
   if (!mongoose.Type.ObjectId.isValid(id))
     return res.status(404).send("Now Memory with this Id");
 
   const memory = await Memory.findById(id);
-  const updatedMemory = await Memory.findByIdAndRemove(
-    id,
-    { likeCount: memory.likeCount + 1 },
-    { new: true }
-  );
+
+  // Check if the userId in the list of likeUser
+  const index = memory.likes.findIndex((id) => id === String(req.userId));
+
+  // if the id doesn't exist
+  if (index === -1) {
+    // like the post
+    memory.likes.push(req.userId);
+  } else {
+    // disLike the post
+    memory.likes = memory.likes.filter((id) => id !== String(req.userId));
+  }
+  const updatedMemory = await Memory.findByIdAndRemove(id, memory, {
+    new: true,
+  });
 
   res.json(updatedMemory);
 };
